hifive-vp: Ensure that GpioServer::quit terminates the receive loop
Without this commit the receive loop would not always be terminated as
it may be stuck in a blocking accept() system call and therefore never
checked the stop class variable.

To make sure that the accept() system call does actually return (and
therefore causes the receive loop to terminate) this commit attempts to
establish a new connection to the server socket after setting the stop
variable.